Rename Command to BaseCommand.
Command will eventually become an interface when all usages have been replaced
